+ Fixed quite a few issues related to ActorComponents being attached to the GameActorProxy, not the GameActor. + As a test, reworked a few simple actors so that they don't do their simulation code in the drawable. + Fixed a few bugs related to 64bit builds.  There are no more known. + Removed the old NECCBoatActor.  it didn't build anyway. + Fixed a fair few deprecation warnings.

+++ source/SimCore/HLA/HLACustomParameterTranslator.cpp
@@ -654,7 +654,7 @@ namespace SimCore
          if (parameterDataType == dtCore::DataType::DOUBLE)
          {
             const dtGame::DoubleMessageParameter& dmp = static_cast<const dtGame::DoubleMessageParameter&>(parameter);
-            ds << (unsigned long)(dmp.GetValue() * 1000.0);
+            ds << (unsigned)(dmp.GetValue() * 1000.0);
          }
       }
